Move components folder
- Move folder to root
