+++ components/background.tsx
