+++ components/content.tsx
