+++ pages/404.tsx
@@ -1,16 +1,16 @@
-import type { NextPage } from 'next'
-import Content from './components/content';
+import type { NextPage } from "next";
+import Content from "../components/content";
 
 const error = "Error 404";
 const message = "This page could not be found.";
 
 const NotFound: NextPage = () => {
-    return (
-        <Content>
-            <p className="text-base mb-4 text-nord11">{error}</p>
-            <p className="text-base mb-4 text-nord6">{message}</p>
-        </Content>
-    )
-}
+  return (
+    <Content>
+      <p className="text-base mb-4 text-nord11">{error}</p>
+      <p className="text-base mb-4 text-nord6">{message}</p>
+    </Content>
+  );
+};
 
 export default NotFound;
