+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from "next/app";
 import "../styles/global.css";
-import Layout from "./components/layout";
+import Layout from "../components/layout";
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
